Move age.js tests to their own file
