+++ root/static/scripts/tests/index-web.js
@@ -22,5 +22,6 @@ require('./release-editor/trackParser');
 require('./release-editor/utils');
 require('./release-editor/validation');
 require('./utility');
+require('./utility/age');
 require('./utility/dates');
 require('./utility/sanitizedEditor');
